pyproject.toml
[tool.poetry]
name = "LED-Matrix"
version = "0.0.0"
description = "APA102 based LED-Matrix controlled by Raspberry Pi on Alpine Linux written in Python."
readme = "README.md"
license = "GPL-3.0-only"
authors = ["Marc Ammon <marc.ammon@hotmail.de>"]
homepage = "https://github.com/mammo0/LED-Matrix"
packages = [
    { include = "led_matrix" }
]
include = [
    { path = "led_matrix/resources", format = ["sdist", "wheel"] }
]


[tool.poetry.scripts]
create-config = "led_matrix.config:create_initial_config"
led-matrix = "led_matrix.cli:run"


[tool.poetry.dependencies]
python = "~3.11"
spidev = { version = "^3.6", markers = "sys_platform == 'linux'"}
imageio = "^2.31.5"
freetype-py = {git = "https://github.com/mammo0/freetype-py.git", branch = "led-matrix"}
simple-plugin-loader = "^3.0.0"

[tool.poetry.group.alpine.dependencies]
numpy = "=1.24.4"
pillow = "=9.5.0"
bottle = "=0.12.25"
apscheduler = "=3.10.1"
tzlocal = "=4.3"

[tool.poetry.group.config.dependencies]
tzlocal = "=4.3"
python-ini = "^1.1.0"
jsons = "^1.6.3"
astral = "^3.2"

[tool.poetry.group.dev.dependencies]
pygame = "^2.5.2"


[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
format-jinja = """
    {%- if distance == 0 -%}
        {{ serialize_pep440(base, stage, revision) }}
    {%- else -%}
        {{ serialize_pep440(base, stage, revision , post=distance) }}
    {%- endif -%}
"""
metadata = false


[tool.pylint.MAIN]
extension-pkg-allow-list = [
    "spidev",
    "pygame"
]


[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
